Finished to look exactly as in video. Next to modify it a little.
Because the video is old, some code is not compatible or deprecated. Like the keyCode part. If used in an addEventListener with function(e) directly; the function keycode will be deprecated. But if used in another independent function and use it in the addEventListener it will work.

The last part of removing the transition was hard. NOt even copying the code from the video or the answer worked for me. Luckily found a comment with a better solution.